src/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.2
daphne==3.0.2
Django==4.0.10
django-celery-beat==2.5.0
django-jazzmin==2.6.0
django-solo==2.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
flake8==7.0.0
frozenlist==1.4.1
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
install==1.3.5
isort==5.13.2
kombu==5.3.5
loguru==0.7.2
mccabe==0.7.0
msgpack==1.0.7
multidict==6.0.4
openpyxl==3.0.10
packaging==23.2
phonenumbers==8.13.29
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pyOpenSSL==24.0.0
pyTelegramBotAPI==4.15.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.4
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zope.interface==6.1
